sending pids on result 4, pid sending clean up, stored trip ids are reset on trip flag 9, getting latest trip id when post trip fails, app gets things from local db when no connection, pids and dtcs are sent when internet connection is resumed

+++ app/src/main/java/com/pitstop/MainActivity.java
@@ -1191,10 +1191,14 @@ public class MainActivity extends AppCompatActivity implements ObdManager.IBluet
     }
 
     private void refreshFromServer() {
-        carIssueList.clear();
-        carLocalStore.deleteAllCars();
-        carIssueLocalStore.deleteAllCarIssues();
-        getCarDetails();
+        if(!NetworkHelper.isConnected(this)) {
+            refreshFromLocal();
+        } else {
+            carIssueList.clear();
+            carLocalStore.deleteAllCars();
+            carIssueLocalStore.deleteAllCarIssues();
+            getCarDetails();
+        }
     }
 
     private void refreshFromLocal() {
